smarthelmet_fixed_upgrade/Flask_HTML_Version/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
python-socketio==5.9.0
python-engineio==4.7.1
paho-mqtt==1.6.1
python-dotenv==1.0.0
pymongo==4.5.0
eventlet==0.33.3
